+++ frontend/manager/base.js
@@ -12,24 +12,14 @@ window.Bokeh = Bokeh;
 
 import DeepZoomImage from '../components/DeepZoomImage.vue';
 
-import Basket from './Basket.vue'
-import SearchResults from './SearchResults.vue';
+import Basket from '../base/Basket.vue'
+import DatasetList from './DatasetList.vue';
 
 import 'topobank/scss/custom.scss';
 /**
  * Event bus for initiating DZI download
  */
-import mitt from 'mitt';
 import axios from "axios";
-
-const eventHub = mitt();
-
-/**
- * Return event hub
- */
-export function getEventHub() {
-    return eventHub;
-}
 
 /**
  * Wrapper for an OpenSeadragon instance (with a scale bar)
@@ -39,7 +29,6 @@ export function createDeepZoomImage(el, csrfToken, props) {
     app.use(createBootstrap());
     axios.defaults.headers.common['X-CSRFToken'] = csrfToken;
     app.provide('csrfToken', csrfToken);
-    app.provide('eventHub', eventHub);
     app.mount(el);
     return app;
 }
@@ -47,12 +36,11 @@ export function createDeepZoomImage(el, csrfToken, props) {
 /**
  * Used to display search results/list of digital surface twins
  */
-export function createSearchResultsApp(el, csrfToken, props) {
-    let app = createApp(SearchResults, props);
+export function createDatasetListApp(el, csrfToken, props) {
+    let app = createApp(DatasetList, props);
     app.use(createBootstrap());
     axios.defaults.headers.common['X-CSRFToken'] = csrfToken;
     app.provide('csrfToken', csrfToken);
-    app.provide('eventHub', eventHub);
     app.mount(el);
     return app;
 }